# Faster LFU Cache 