In response to #260 this PR adds a faster implementation of the `LFUCache` to `cachetools`. ~~I aimed to minimise the changes as much as possible but it does effectively need to throw out most of the base `Cache` classes work unfortunately, the dunder private methods make accessing the parents attributes directly messy. Thus it still subclasses `Cache` but re-implements all of its methods. I am open to opinions on other possible implementations?~~

# Implementation Notes

This isn't an O(1) implementation as was discussed in the original issue as it updates the minimum frequency via `min` over the buckets (bounded on distinct frequency buckets which is <= cache size), the `OrderedDict` gives LRU tie-breaking within each frequency bucket. I think this gives a nice balanced of implementation complexity and performance.

# Benchmarking

## [`cacheing` code](https://github.com/breid48/cacheing/blob/main/benchmark/benchmark.py)
I ran the benchmarking code from `cacheing` which was mentioned in #260 and we can see the performance is now much closer between the two (tested on M1 Macbook Pro):
```bash
$ python ./benchmark.py --cache cachetools.LFUCache cacheing.LFUCache --method set get delete

Set: 
       library cache_type  cache_size    median       p90       p99
0  cachetools   LFUCache         256  1.458044  1.540990  1.666952
1  cachetools   LFUCache        1024  1.500011  1.584005  3.708422
2  cachetools   LFUCache        4096  1.459033  1.541979  1.665961
3  cachetools   LFUCache       16384  0.457978  1.500011  1.625041
4    cacheing   LFUCache         256  0.667002  0.709028  0.833999
5    cacheing   LFUCache        1024  0.624976  0.667002  0.708038
6    cacheing   LFUCache        4096  0.624976  0.667002  0.707980
7    cacheing   LFUCache       16384  0.333006  0.625034  0.708038
Get: 
       library cache_type  cache_size    median       p90       p99
0  cachetools   LFUCache         256  0.416010  0.500004  0.707980
1  cachetools   LFUCache        1024  0.457978  0.500004  1.417007
2  cachetools   LFUCache        4096  0.459026  0.540982  0.625034
3  cachetools   LFUCache       16384  0.458967  0.540982  0.583997
4    cacheing   LFUCache         256  0.333006  0.416010  0.666012
5    cacheing   LFUCache        1024  0.374974  0.416010  0.917003
6    cacheing   LFUCache        4096  0.375032  0.457978  0.541040
7    cacheing   LFUCache       16384  0.374974  0.457978  0.667002
Delete: 
       library cache_type  cache_size    median       p90       p99
0  cachetools   LFUCache         256  0.292028  0.957982  1.166429
1  cachetools   LFUCache        1024  0.333006  0.375032  0.999994
2  cachetools   LFUCache        4096  0.333006  0.375032  0.502053
3  cachetools   LFUCache       16384  0.333006  0.375032  0.417000
4    cacheing   LFUCache         256  0.250002  0.291970  0.473263
5    cacheing   LFUCache        1024  0.291038  0.333006  0.374974
6    cacheing   LFUCache        4096  0.291970  0.333996  0.417000
7    cacheing   LFUCache       16384  0.291970  0.333006  0.375032
```

## Direct Comparison
I also wrote a small script to side by side the cache with the original implementation, `LFUCacheOld` is the current implementation on `master`.

__Code__
```python
import time
import random
from cachetools import LFUCache, LFUCacheOld


def benchmark(cache_cls, capacity=4_096, operations=500_000, key_space=8_192):
    rnd = random.Random(0)
    cache = cache_cls(maxsize=capacity)

    start = time.perf_counter()

    for _ in range(operations):
        k = rnd.randrange(key_space)

        if k in cache:
            _ = cache[k]
        else:
            cache[k] = k

    elapsed = time.perf_counter() - start
    print(f"{cache_cls.__name__:>11}: {elapsed:6.2f}s | ({operations / elapsed:,.0f} ops/s)")


def main():
    benchmark(LFUCacheOld)
    benchmark(LFUCache)


if __name__ == "__main__":
    main()
```

__Results__
```bash
LFUCacheOld:  17.54s | (28,503 ops/s)
   LFUCache:   0.71s | (700,387 ops/s)
```

## Using the original gist
The LFU is much closer to LRU now.
```bash
$ python3 -m timeit -s'import random; import cachetools; c = cachetools.LRUCache(1000)' 'i = random.randint(1, 10000); c[i] = i'
200000 loops, best of 5: 1.22 usec per loop
$ python3 -m timeit -s'import random; import cachetools; c = cachetools.LFUCache(1000)' 'i = random.randint(1, 10000); c[i] = i'
200000 loops, best of 5: 1.77 usec per loop
```