@tkem 2025-06-12
@breid48: Thanks for the response, and thanks again for pointing me at

Ketan Shah, Anirban Mitra, and Dhruv Matani, An O(1) algorithm for implementing the LFU cache eviction scheme, (August 16, 2010).

So, credit where it's due ;-)